compensatory leave part 2

+++ Models/companleaveBal.cs
@@ -17,7 +17,7 @@ namespace HRworks.Models
         public int Id { get; set; }
         public int EmpNo { get; set; }
         public System.DateTime dateModified { get; set; }
-        public Nullable<long> balance { get; set; }
+        public Nullable<double> balance { get; set; }
     
         public virtual master_file master_file { get; set; }
     }
